Add doxconf crossref aliases \icdcmd \icdpubev \icdalarm \icdsubev
We can now crosslink all of the major items from model files.

In the case of alarms and subscribed events it's not clear whether
the links take you to the correct line of the table, or just the
page with the table. In either case the link is still useful.

For commands, alarms, and published events the argument of the alias
is just the last element of the fully-qualified name (e.g., "foo" in the
case of "component.foo".

However, for subscribed events you specify the fully-qualified name in the
argument (e.g., "component.foo").