latex: update header.tex in an attempt with work with doxygen 1.8.17
After upgrading to Ubuntu 20.04 I found that documents were failing
to build, with LaTex complaining about \DoxyHorRuler . I eventually
tracked it down to the fact that we are defining custom headers, and
failing to pick up some modifications present in the default header
created by Doxygen 1.8.17. With this patch documents appear to
build on my Ubuntu 20.04 system, but testing will be required on
other systems to confirm.